requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
aiosmtplib==4.0.0
redis==5.0.1
python-dotenv==1.0.1
pytest==8.0.1
httpx==0.26.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0 